+++ backend/controllers/profiles.js
@@ -66,6 +66,7 @@ router.get('/:id', (req, res) => {
     .then(profile => {
         // console.log(profile.characters)
         res.render('profiles/show', { profile })
+        console.log(profile.characters)
     })
     // catch any error
     .catch(err => {
@@ -178,9 +179,10 @@ router.get('/:id/characters/new', (req, res) => {
 })
 
 // SHOW SINGLE CHARACTER PAGE
-router.get('/:id/characters/:id', (req, res) => {
-    res.send('GET /profile/:id/characters/:id')
-})
+// router.get('/:id/characters/:id', (req, res) => {
+//     // res.send('GET /profile/:id/characters/:id')
+//     res.render('profiles/characters/show')
+// })
 
 // EDIT CHARACTER
 router.put('/:id/characters/:id', (req, res) => {